feat: implement medication detail navigation and refine UI components
- Integrated `feature:detail` module into the application navigation and `bio-scan` feature.
- Implemented `MedicationDetailScreen` to display comprehensive medicine information, including dosage, validity, cures, and side effects.
- Updated `MedicationCardItemData` and `MedicationDetailNavKey` to use `Long` for IDs to ensure consistency with data models.
- Refined typography by reducing line height for body styles in the design system.
- Improved `MedicationCardList` to support click callbacks for navigating to medicine details.
- Enhanced `MedicationEditViewModel` to correctly handle medication types, salts, and instructions during the saving process.
- Fixed date formatting logic to use epoch seconds instead of milliseconds.
- Adjusted UI layouts in detail screens, including better spacing for flow rows and improved alignment for bullet points.
- Registered `MedicationDetailEntry` in the main app navigation graph.

+++ core/design-system/src/main/kotlin/medyo/com/core/design_system/component/card/MedicineCardItem.kt
@@ -193,7 +193,7 @@ private fun MedicationCardItemPreview() {
 }
 
 data class MedicationCardItemData(
-    val id: Int, // Important for LazyColumn keys
+    val id: Long, // Important for LazyColumn keys
     val name: String,
     val dosage: String,
     val iconRes: Int,
