+++ feature/detail/api/src/main/kotlin/medyo/com/feature/detail/api/MedicationDetailNavKey.kt
@@ -5,4 +5,4 @@ import androidx.navigation3.runtime.NavKey
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class MedicationDetailNavKey(val medicationId: Int) : NavKey
+data class MedicationDetailNavKey(val medicationId: Long) : NavKey
